* commit '004a88dda93fe8756839dfeb391832edc2870fe4': (40 commits)
  Implement version of step() that takes a scalar edge and a vector for the other argument.
  Revert "Revert "Only build the tests for top-level builds (i.e. not during mm/mmm for driver).""
  Revert "Only build the tests for top-level builds (i.e. not during mm/mmm for driver)."
  Only build the tests for top-level builds (i.e. not during mm/mmm for driver).
  Return errors to user when copy fails.
  One more set of libraries to remove for cleaning up after the header rename.
  Remove old library builds with CleanSpec too.
  update cleanspec for header generator
  Switch to generated headers.
  Update runtime header generator
  Make update_rs_prebuilts.sh script work properly on Darwin.
  Make object destruction asynchronous.
  Add handcode mode to rsg_generator.
  Remove unnecessary Android.mk.
  Fix HelloComputeNDK for RS NDK API update.
  Add user-defined cache path.
  Runtime spec file and generation
  Remove excessive ALOGE.
  Fix memory leak due to smart pointer misuse.
  implement finish
  ...

+++ rs.h
@@ -45,10 +45,10 @@ extern "C" {
     // Allocation update
     const void* rsaAllocationGetType(RsContext con, RsAllocation va);
     // Type update
-    void rsaTypeGetNativeData(RsContext, RsType, uint32_t *typeData, uint32_t typeDataSize);
+    void rsaTypeGetNativeData(RsContext, RsType, uintptr_t *typeData, uint32_t typeDataSize);
     // Element update
-    void rsaElementGetNativeData(RsContext, RsElement, uint32_t *elemData, uint32_t elemDataSize);
-    void rsaElementGetSubElements(RsContext, RsElement, uint32_t *ids, const char **names,
+    void rsaElementGetNativeData(RsContext, RsElement, uintptr_t *elemData, uint32_t elemDataSize);
+    void rsaElementGetSubElements(RsContext, RsElement, uintptr_t *ids, const char **names,
                                   uint32_t *arraySizes, uint32_t dataSize);
 
     RsDevice rsDeviceCreate();
